+++ src/gui.cpp
@@ -1107,12 +1107,15 @@ const std::vector<std::string>& GUI::occt_material_combo_labels_()
 }
 
 void GUI::sketch_list_inspector_(const Sketch::sptr& sketch, int index, Sketch::sptr& hover_dim_sketch,
-                                 size_t& hover_dim_index, Sketch::sptr& hover_face_sketch, size_t& hover_face_index)
+                                 size_t& hover_dim_index, Sketch::sptr& hover_face_sketch, size_t& hover_face_index,
+                                 Sketch::sptr& hover_edge_sketch, size_t& hover_edge_index,
+                                 Sketch::sptr& hover_node_sketch, size_t& hover_node_index)
 {
   ImGui::Indent();
   ImGui::PushID(index);
 
-  const auto draw_section = [](const char* title, const std::vector<std::string>& labels)
+  const auto draw_hover_section = [&](const char* title, const std::vector<std::string>& labels,
+                                      Sketch::sptr& hover_sketch, size_t& hover_index)
   {
     const size_t       count = labels.size();
     ImGuiTreeNodeFlags flags = ImGuiTreeNodeFlags_SpanAvailWidth;
@@ -1121,8 +1124,17 @@ void GUI::sketch_list_inspector_(const Sketch::sptr& sketch, int index, Sketch::
 
     if (ImGui::TreeNodeEx(title, flags, "%s (%zu)", title, count))
     {
-      for (const std::string& label : labels)
-        ImGui::BulletText("%s", label.c_str());
+      for (size_t i = 0; i < count; ++i)
+      {
+        ImGui::PushID(static_cast<int>(i));
+        ImGui::BulletText("%s", labels[i].c_str());
+        if (ImGui::IsItemHovered() && sketch->is_visible())
+        {
+          hover_sketch = sketch;
+          hover_index  = i;
+        }
+        ImGui::PopID();
+      }
 
       if (count > 0)
         ImGui::TreePop();
@@ -1197,8 +1209,8 @@ void GUI::sketch_list_inspector_(const Sketch::sptr& sketch, int index, Sketch::
     }
   }
 
-  draw_section("Nodes", sketch->inspector_node_labels());
-  draw_section("Edges", sketch->inspector_edge_labels());
+  draw_hover_section("Nodes", sketch->inspector_node_labels(), hover_node_sketch, hover_node_index);
+  draw_hover_section("Edges", sketch->inspector_edge_labels(), hover_edge_sketch, hover_edge_index);
 
   {
     const std::vector<std::string> labels = sketch->inspector_face_labels();
@@ -1266,7 +1278,9 @@ void GUI::sketch_list_()
   {
     m_view->set_sketch_list_hover(nullptr);
     m_view->set_sketch_list_measurement_hover(nullptr, SIZE_MAX);
-    m_view->set_sketch_list_face_hover(nullptr, SIZE_MAX);
+    m_view->set_sketch_list_hover_face(nullptr, SIZE_MAX);
+    m_view->set_sketch_list_hover_edge(nullptr, SIZE_MAX);
+    m_view->set_sketch_list_hover_node(nullptr, SIZE_MAX);
     return;
   }
 
@@ -1283,7 +1297,9 @@ void GUI::sketch_list_()
   {
     m_view->set_sketch_list_hover(nullptr);
     m_view->set_sketch_list_measurement_hover(nullptr, SIZE_MAX);
-    m_view->set_sketch_list_face_hover(nullptr, SIZE_MAX);
+    m_view->set_sketch_list_hover_face(nullptr, SIZE_MAX);
+    m_view->set_sketch_list_hover_edge(nullptr, SIZE_MAX);
+    m_view->set_sketch_list_hover_node(nullptr, SIZE_MAX);
     ImGui::End();
     return;
   }
@@ -1296,8 +1312,12 @@ void GUI::sketch_list_()
   Sketch::sptr sketch_list_hover;
   Sketch::sptr sketch_list_measurement_hover_sketch;
   size_t       sketch_list_measurement_hover_index = SIZE_MAX;
-  Sketch::sptr sketch_list_face_hover_sketch;
-  size_t       sketch_list_face_hover_index = SIZE_MAX;
+  Sketch::sptr sketch_list_hover_face_sketch;
+  size_t       sketch_list_hover_face_index = SIZE_MAX;
+  Sketch::sptr sketch_list_hover_edge_sketch;
+  size_t       sketch_list_hover_edge_index = SIZE_MAX;
+  Sketch::sptr sketch_list_hover_node_sketch;
+  size_t       sketch_list_hover_node_index = SIZE_MAX;
   for (Sketch::sptr& sketch : m_view->get_sketches())
   {
     EZY_ASSERT(sketch);
@@ -1443,14 +1463,17 @@ void GUI::sketch_list_()
 
     if (expanded && ui_show_sketch_list_expand())
       sketch_list_inspector_(sketch, index, sketch_list_measurement_hover_sketch, sketch_list_measurement_hover_index,
-                             sketch_list_face_hover_sketch, sketch_list_face_hover_index);
+                             sketch_list_hover_face_sketch, sketch_list_hover_face_index, sketch_list_hover_edge_sketch,
+                             sketch_list_hover_edge_index, sketch_list_hover_node_sketch, sketch_list_hover_node_index);
 
     ++index;
   }
 
   m_view->set_sketch_list_hover(sketch_list_hover);
   m_view->set_sketch_list_measurement_hover(sketch_list_measurement_hover_sketch, sketch_list_measurement_hover_index);
-  m_view->set_sketch_list_face_hover(sketch_list_face_hover_sketch, sketch_list_face_hover_index);
+  m_view->set_sketch_list_hover_face(sketch_list_hover_face_sketch, sketch_list_hover_face_index);
+  m_view->set_sketch_list_hover_edge(sketch_list_hover_edge_sketch, sketch_list_hover_edge_index);
+  m_view->set_sketch_list_hover_node(sketch_list_hover_node_sketch, sketch_list_hover_node_index);
 
   ImGui::EndChild();